Binary files needed to bootstrap TENEX.
- dtboot.bin are DECtape boot blocks.
- tenex.sav is the resident monitor.
- tenex.swp is the swappable monitor.
- dluser.sav creates directories.
- users.txt specified directory data.
- dumper.sav restores files from magtape.

Dtboot.bin is to make a bootable DECtape.  Tenex.sav and tenex.swp
should go on the DECtape.  Note that the resident monitor has TENDMP
built in for loading the swappable monitor.  TENDMP can only read
files with increasing block numbers.

Dluser.sav, users.txt, and dumper.sav can go on a DECtape or on
magtape.  DLUSER can be run from the mini-exec with users.txt as
input, and after that DUMPER can be run.

The media.sh script makes DECtape and magtape images.